inventaireAI/requirements.txt
google-generativeai
pandas
python-dotenv
Pillow
pyinstaller
rembg
onnxruntime